requirements.txt
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
Pillow==10.1.0
Pygments==2.17.2
rich==13.7.0
shellingham==1.5.4
soupsieve==2.5
typer==0.9.0
typing_extensions==4.8.0
yarl==1.9.3
